I just got the flower details and style it to some level
