+++ frontend/package.json
@@ -1,4 +1,4 @@
-{
+{ "proxy": "http://localhost:7000",
   "name": "frontend",
   "version": "0.1.0",
   "private": true,
